feat: move user dto
