+++ lib/users/users.ts
@@ -1,6 +1,10 @@
 import { auth0Fetch } from '@/lib/auth/fetch';
 import { managementApiToken } from '@/lib/auth/management-api-token';
+import { USER_FIELDS, type UserDto, type UserListDto, type UserListQuery } from '@/lib/users/user.dto';
 import { FetchError } from '../utils/fetch';
+
+// Constants
+const USER_CACHE_TIMEOUT = 60; // s
 
 // Api calls
 export async function queryUsers(query: UserListQuery & { includeTotals: true }): Promise<UserListDto>;
@@ -9,7 +13,7 @@ export async function queryUsers(query?: UserListQuery): Promise<UserListDto | U
 export async function queryUsers(query: UserListQuery = {}): Promise<UserListDto | UserDto[]> {
   console.log('[auth0] Load users');
   const url = new URL(`https://${process.env.AUTH0_DOMAIN}/api/v2/users`);
-  url.searchParams.set('fields', 'user_id,name,nickname,picture,identities,last_login,app_metadata');
+  url.searchParams.set('fields', USER_FIELDS.join(','));
   url.searchParams.set('include_totals', query.includeTotals ? 'true' : 'false');
 
   if (query.page !== undefined) {
@@ -26,8 +30,8 @@ export async function queryUsers(query: UserListQuery = {}): Promise<UserListDto
       Authorization: `Bearer ${await managementApiToken()}`
     },
     next: {
-      revalidate: 60,
-      tags: ['users']
+      revalidate: USER_CACHE_TIMEOUT,
+      tags: ['users', 'users-pages']
     }
   });
 }
@@ -35,7 +39,7 @@ export async function queryUsers(query: UserListQuery = {}): Promise<UserListDto
 export async function queryUser(id: string): Promise<UserDto | null> {
   console.log(`[auth0] Load user "${id}"`);
   const url = new URL(`https://${process.env.AUTH0_DOMAIN}/api/v2/users/${id}`);
-  url.searchParams.set('fields', 'user_id,name,nickname,picture,identities,last_login,app_metadata');
+  url.searchParams.set('fields', USER_FIELDS.join(','));
 
   try {
     return await auth0Fetch<UserDto>(url, {
@@ -44,7 +48,7 @@ export async function queryUser(id: string): Promise<UserDto | null> {
         Authorization: `Bearer ${await managementApiToken()}`
       },
       next: {
-        revalidate: 60,
+        revalidate: USER_CACHE_TIMEOUT,
         tags: ['users', `users-${id}`]
       }
     });
@@ -57,34 +61,3 @@ export async function queryUser(id: string): Promise<UserDto | null> {
   }
 }
 
-// Types
-export interface UserDto {
-  readonly name: string;
-  readonly nickname: string;
-  readonly identities: UserIdentityDto[],
-  readonly user_id: string;
-  readonly picture: string;
-  readonly last_login: string;
-  readonly app_metadata: Record<string, unknown>;
-}
-
-export interface UserIdentityDto {
-  readonly connection: string;
-  readonly user_id: string;
-  readonly provider: string;
-  readonly isSocial: boolean;
-}
-
-export interface UserListDto {
-  readonly start: number;
-  readonly limit: number;
-  readonly length: number;
-  readonly total: number;
-  readonly users: UserDto[];
-}
-
-export interface UserListQuery {
-  readonly includeTotals?: boolean;
-  readonly page?: number;
-  readonly perPage?: number;
-}
